This commit was generated by cvs2svn to compensate for changes in r6803, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@6804 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/aciacore.c
@@ -242,7 +242,7 @@ int myacia_write_snapshot_module(snapshot_t * p)
     snapshot_module_write_byte(m, (BYTE)(status | (irq?0x80:0)));
     snapshot_module_write_byte(m, cmd);
     snapshot_module_write_byte(m, ctrl);
-    snapshot_module_write_byte(m, intx);
+    snapshot_module_write_byte(m, (BYTE)(intx));
 
     if(alarm_active) {
         snapshot_module_write_dword(m, (acia_alarm_clk - myclk));
